Add Portainer stack (#8)
Domain defaults to localhost which can be used when testing. Otherwise
set to your own domain.  You can also set the admin user password which
defaults to "admin". Note that this cannot  be used for changing the admin
password later.

If you use too easy password Portainer will force you to change it
when you log in for the first time.

$ export INITIAL_PASSWORD=$(htpasswd -nB admin | cut -d ":" -f 2)
$ DOMAIN=example.com docker stack deploy -c portainer.yaml portainer

https://portainer.example.com